Add tags support for Import Recipe
This replaces the fake data from before with calls to the backend to
get tag suggestions from the database.  When a suggestion is chosen,
the tag information is added to the Redux store, which is then
accessed to pass that information to the backend.

TODO: clear tags from UI when they're cleared from the Redux store