Don't use Redux for recipe import
Redux really isn't necessary for this operation, but writing the code
was a good way to learn Redux.  Now that Redux is being used for other
things, switch to just using direct Ajax calls.

This also slightly restructures the backend to correctly use POST
parameters instead of query parameters when invoking the POST endpoint
to import a recipe.