Fix issue with recipe import errors on server not propagating
Rearrange the recipe import code so that instead of calling the
asynchronous recipe creation method within a then block, we call it
outside of that block using await.  This allows errors returned by
the recipe creation method to be correctly propagated to the caller.

The frontend code is also modified to display information on errors
returned by recipe creation, and to account for the fact that recipe
import on the server now returns the newly created record.