Connect Favorite and Meal Planner buttons on Add tab to database
The database schema is modified so that each entry in the saved_recipes
table has a 'value' column indicating whether the entry represents a
true or a false value, and the server model is updated accordingly.
The client code checks the database when the component is mounted
to set the state appropriately, and updates the state in the database
when the respective icons are clicked.