Connect RecipeTable and SortablePaginatedTable to database
Redux is used to load data from the database as needed, and serve the
data from the store if the data hasn't changed between requests.

The ReviewTable class still needs to be integrated with this new
framework: that will happen in a separate commit.

This also includes an unrelated change in one of the model classes
to rename a variable.