Connect Plan and Favorites tabs to database
This defines new actions and reducers to handle these tabs, and refactors
the action creators for fetching recipes to handle the saved recipe case
as well.  The backend is also modified to add a new parameter to recipe
lookup that allows only favorites or meal planner recipes to be returned,
assuming a default user ID.

A bug with navigating between the Search, Plan, and Favorites tabs
is also fixed.