Implement selection and deletion in recipe tables
This adds a modal dialog to prompt for confirmation of the deletion.
Logic is also added to reload from the server after deletion, including
methods to clear the Redux cache and force a reload from the database
rather than from cache.

The method to clear the Redux cache has also been added to the Import
Recipe functionality, so going to the Search page after doing an import
reloads data from the server (which needs to take the newly imported
recipe into account).